Visualization#3
I chose this visual to highlight the difference in the two scores of schools within the 3-year timespan. I chose Roy H Mann because that is my alma mater. I selected East Side Middle school not only because of the high scores but also that would be my son's Zone school since we live in the area of East Side Middle.